Rebuild the datahub containers to pick up new base JDK
The underlying Java runtime container on which datahub is built has ben
upgraded from 11.0.9 to 11.0.16. This noop change triggers a rebuild of
the datahub containers in order to pick up this new base image.

We are hopeful that this will fix an issue we have with deploying
datahub to the staging-codfw kubernetes cluster.

Bug: T327799
Change-Id: Ib9df6382af09b4772947f802760a72eee815c608