Add a datahub-upgrade container
We are adding a new container to the pipelinelib configuration for
datahub, namely datahub-upgrade. This container includes only a jar file
that is compiled using gradle and is useful for a number of tasks such
as performing makor version upgrades and recreating the elasticsearch
indices.

The JDK version for the kafka-setup container is also updated in this
commit due to its being outdated and causing a build error.

Bug: T329514
Change-Id: Ia1f0e06b0fdb1fc6b3e8755744e6b3fc53cee5fd