Update the kafka-setup conainer of datahub
We have previously not used the kafka-setup container, but had setup
kafka by-hand for our instances. Now we would like to start using it
because it is the mechanism that upstream uses to introduce changes to
the kafka topic handling in Datahub.

This commit updates our image to bring it more into line with the
upstream version. Specifically, it copies in two more scripts to the
container and changes a few of the environment variables relating top
topic names.

Bug: T329514
Change-Id: I21abbea3637b630b1baaca492f84fc9e3bd09bb7