Improve the datahub kafka-setup container
The kafka-setup container required considerable work in order to get it
to the point where it was usable with our configuration.

I have taken the opportunity to start to de-fork it, so the kafka-setup
container here downloads the datahub source, rather than using the local
copy.

The blubber pipeline also builds the required 'dockerize' and 'cub'
commands, which were steps we previously carried out in our own custom
setup script. The custom setup script has been removed.

Bug: T329514
Change-Id: Ib85637a0bde2111c9f73249749a199de820acb76