Fix the kafka-setup container for datahub
The kafka-setup container still didn't work because it tried to rely on
bash aliases to make various commands available.

This commit fixes that by supplying small scripts for kafka-topics.sh
and kafka-configs.sh that point to the versions without the .sh suffix.

In addition to that, we add /datahub to the path of the container.

Bug: T329514
Change-Id: I7e52d59896338c661591c5f6959816de867ea536