Begin un-forking datahub from the upstream
This change to the blubber configuration files is intended to allow us
to un-fork our code from the upstream sources, since instead of copying
the datahub source into each build container it is shallow cloned from
GitHub instead.

The intention is that we keep our build pipeline separate and ensure
that we can run it against unmodified upstream sources.

Bug: T329514
Bug: T303381
Change-Id: I2302224dc940c5c3f5081499321ab22a4d69cfe9